Calculating Percentage Changes

pctchg takes a dataframe and a numeric column in it to calculate percentage changes from one period to the next. If values are missing/0, the function will return 0 for that observation in the "pctchg_" column. Time series dataframes recommended.

Usage

pctchg(df, col)

Arguments

df

dataframe containing column of interest

col

name of the column, passed in quotes

Value

a new dataframe with an additional column containing "pctchg_"

Examples

pctchg(treedat, 'x')
#>     period         x           y z    pctchg_x
#> 1        1  1.048530    5.916203 a          NA
#> 2        2  1.021098    7.444445 b  -2.6162181
#> 3        3  1.187908    2.437572 b  16.3363558
#> 4        4  1.324990    8.075252 a  11.5397551
#> 5        5  1.270861    4.735745 a  -4.0852556
#> 6        6  1.322699    6.522545 b   4.0789630
#> 7        7  1.203185    8.866601 a  -9.0355950
#> 8        8  1.393374   16.015974 a  15.8071659
#> 9        9  1.315332    2.530172 b  -5.6009473
#> 10      10  1.555536    9.018722 b  18.2618227
#> 11      11  1.418565    8.634356 b  -8.8053728
#> 12      12  1.304490   14.918610 b  -8.0416236
#> 13      13  1.439100   21.651528 b  10.3190242
#> 14      14  1.589373   15.665018 b  10.4421318
#> 15      15  1.405687   10.634927 b -11.5571322
#> 16      16  1.661519   10.604511 a  18.1997967
#> 17      17  1.589126   14.638448 a  -4.3570663
#> 18      18  1.649037    7.766527 b   3.7701121
#> 19      19  1.675417   13.873119 b   1.5996688
#> 20      20  2.047280   17.363223 a  22.1952706
#> 21      21  2.024222    9.923722 b  -1.1262886
#> 22      22  1.838700   19.239874 b  -9.1650962
#> 23      23  2.140325   20.134298 a  16.4042779
#> 24      24  1.849779   13.730836 b -13.5748428
#> 25      25  2.018241   17.628260 b   9.1071061
#> 26      26  2.181264   12.113236 b   8.0775149
#> 27      27  2.146566   61.376088 b  -1.5907445
#> 28      28  2.590519   30.506691 b  20.6819865
#> 29      29  2.249139   42.411416 b -13.1780341
#> 30      30  2.496766   31.758915 a  11.0098373
#> 31      31  2.411410   23.494780 a  -3.4186619
#> 32      32  2.634877   44.098945 b   9.2670608
#> 33      33  2.427350   37.552375 b  -7.8761441
#> 34      34  2.478136   17.110185 a   2.0922341
#> 35      35  2.806140   38.842257 b  13.2359379
#> 36      36  2.992099   86.024466 a   6.6268405
#> 37      37  3.413330   80.283497 b  14.0781130
#> 38      38  2.770799   56.492578 b -18.8241539
#> 39      39  3.481619   15.763383 a  25.6539683
#> 40      40  3.390437   62.070485 a  -2.6189666
#> 41      41  3.757685   77.982268 b  10.8318928
#> 42      42  3.973387  113.069851 b   5.7402791
#> 43      43  3.955616   79.780881 b  -0.4472595
#> 44      44  3.425658   77.133263 a -13.3975874
#> 45      45  3.457591   41.819317 b   0.9321457
#> 46      46  3.432784   71.101987 b  -0.7174658
#> 47      47  4.130128   67.913976 a  20.3142635
#> 48      48  4.212303   67.516120 b   1.9896315
#> 49      49  4.554943   60.694355 b   8.1342776
#> 50      50  4.390048  182.277737 b  -3.6201315
#> 51      51  5.173433  226.414247 b  17.8445573
#> 52      52  4.657789   37.544823 b  -9.9671551
#> 53      53  4.096762   49.971455 b -12.0449174
#> 54      54  5.382198  124.201434 b  31.3768823
#> 55      55  5.811703  301.076529 a   7.9801119
#> 56      56  4.969985  108.288521 b -14.4831596
#> 57      57  5.255501   77.484601 a   5.7447998
#> 58      58  5.841287  102.507811 b  11.1461472
#> 59      59  6.406831   55.116357 b   9.6818381
#> 60      60  6.591209  108.978838 b   2.8778395
#> 61      61  6.015107  114.727954 a  -8.7404531
#> 62      62  6.422468  206.109888 a   6.7722856
#> 63      63  6.404346  105.896179 b  -0.2821634
#> 64      64  6.845494  179.799496 b   6.8882576
#> 65      65  8.287669  440.696516 b  21.0675153
#> 66      66  6.556930  143.159905 b -20.8832993
#> 67      67  6.839105  253.297191 a   4.3034566
#> 68      68  6.820978  148.031866 b  -0.2650471
#> 69      69  8.475932  196.747405 a  24.2627039
#> 70      70  7.321359  196.645344 a -13.6217833
#> 71      71  8.983061  425.572213 a  22.6966355
#> 72      72  8.342031  173.885201 a  -7.1359793
#> 73      73  8.829233  431.292281 a   5.8403262
#> 74      74 10.103022  544.570205 b  14.4269490
#> 75      75  9.493691  481.027882 a  -6.0311831
#> 76      76 10.165573  795.033329 b   7.0771469
#> 77      77 11.311990  591.696582 b  11.2774476
#> 78      78  8.350791  214.149361 b -26.1775295
#> 79      79  9.454222  327.620394 b  13.2134919
#> 80      80 10.557685  400.011339 a  11.6716438
#> 81      81 11.370147  351.541415 a   7.6954569
#> 82      82 12.386250  760.670642 a   8.9365865
#> 83      83 14.408895 1042.824597 a  16.3297599
#> 84      84 12.955757  834.535160 a -10.0850048
#> 85      85 13.427992  600.051385 b   3.6449794
#> 86      86 10.973503  701.566971 a -18.2788988
#> 87      87 12.941316  767.197390 a  17.9324086
#> 88      88 13.301346 1586.265187 b   2.7820228
#> 89      89 15.142918 1398.920370 a  13.8450018
#> 90      90 14.693749 1411.897922 b  -2.9661978
#> 91      91 17.621758  684.678216 b  19.9269043
#> 92      92 17.560224 1205.612240 a  -0.3491977
#> 93      93 17.246249 2450.378598 b  -1.7879883
#> 94      94 16.703752  894.650082 b  -3.1455948
#> 95      95 16.625023  648.027046 b  -0.4713233
#> 96      96 17.367347 1731.591270 a   4.4650990
#> 97      97 20.709151 2290.288790 b  19.2418810
#> 98      98 20.784295 3129.811765 a   0.3628549
#> 99      99 20.851572 2292.868697 b   0.3236919
#> 100    100 21.158037  963.157688 a   1.4697426